Split frontend models into modules
