+++ thoth-manager/src/api/contributor_query.rs
@@ -1,7 +1,7 @@
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::api::models::Contributor;
+use crate::api::models::contributor::Contributor;
 
 pub const CONTRIBUTOR_QUERY: &str = "
     query ContributorQuery($contributorId: Uuid!) {
