+++ thoth-manager/src/api/contributors_query.rs
@@ -1,7 +1,7 @@
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::api::models::Contributor;
+use crate::api::models::contributor::Contributor;
 
 pub const CONTRIBUTORS_QUERY: &str = "
     query ContributorsQuery($filter: String) {
