+++ thoth-manager/src/api/work_query.rs
@@ -1,10 +1,10 @@
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::api::models::Imprint;
-use crate::api::models::Work;
-use crate::api::models::WorkStatusDefinition;
-use crate::api::models::WorkTypeDefinition;
+use crate::api::models::imprint::Imprint;
+use crate::api::models::work::Work;
+use crate::api::models::work::WorkStatusDefinition;
+use crate::api::models::work::WorkTypeDefinition;
 
 pub const WORK_QUERY: &str = "
     query WorkQuery($workId: Uuid!) {
