+++ thoth-manager/src/component/contributions_form.rs
@@ -17,9 +17,9 @@ use crate::api::contributors_query::FetchActionContributors;
 use crate::api::contributors_query::FetchContributors;
 use crate::api::contributors_query::Variables;
 use crate::api::contributors_query::CONTRIBUTORS_QUERY;
-use crate::api::models::Contribution;
-use crate::api::models::ContributionTypeValues;
-use crate::api::models::Contributor;
+use crate::api::models::contribution::Contribution;
+use crate::api::models::contribution::ContributionTypeValues;
+use crate::api::models::contributor::Contributor;
 use crate::component::utils::FormBooleanSelect;
 use crate::component::utils::FormContributionTypeSelect;
 use crate::component::utils::FormTextInput;
