+++ thoth-manager/src/component/issues_form.rs
@@ -13,8 +13,8 @@ use crate::api::serieses_query::FetchActionSerieses;
 use crate::api::serieses_query::FetchSerieses;
 use crate::api::serieses_query::Variables;
 use crate::api::serieses_query::SERIESES_QUERY;
-use crate::api::models::Issue;
-use crate::api::models::Series;
+use crate::api::models::issue::Issue;
+use crate::api::models::series::Series;
 use crate::component::utils::FormNumberInput;
 use crate::string::EMPTY_ISSUES;
 use crate::string::REMOVE_BUTTON;
