+++ thoth-manager/src/component/work.rs
@@ -14,13 +14,13 @@ use crate::agent::notification_bus::NotificationBus;
 use crate::agent::notification_bus::NotificationDispatcher;
 use crate::agent::notification_bus::NotificationStatus;
 use crate::agent::notification_bus::Request;
-use crate::api::models::Contribution;
-use crate::api::models::Imprint;
-use crate::api::models::Issue;
-use crate::api::models::Publication;
-use crate::api::models::Work;
-use crate::api::models::WorkStatusValues;
-use crate::api::models::WorkTypeValues;
+use crate::api::models::contribution::Contribution;
+use crate::api::models::imprint::Imprint;
+use crate::api::models::issue::Issue;
+use crate::api::models::publication::Publication;
+use crate::api::models::work::Work;
+use crate::api::models::work::WorkStatusValues;
+use crate::api::models::work::WorkTypeValues;
 use crate::api::work_query::FetchActionWork;
 use crate::api::work_query::FetchWork;
 use crate::api::work_query::Variables;
